Move adding Vineyard and SmellDetails to Wine tests to separate files
